refactor: ViewHistoryId 내부 클래스로 변경
