+++ wnp-server/src/main/java/com/wonnapark/wnpserver/domain/episode/ViewHistory.java
@@ -1,12 +1,17 @@
 package com.wonnapark.wnpserver.domain.episode;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
 
 @Entity
 @Table(name = "view_historys")
@@ -14,15 +19,31 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ViewHistory {
 
+    @Embeddable
+    @Getter
+    @NoArgsConstructor(access = AccessLevel.PROTECTED)
+    @EqualsAndHashCode
+    private static class ViewHistoryId implements Serializable {
+
+        @Column(name = "user_id", nullable = false)
+        private Long userId;
+
+        @Column(name = "episode_id", nullable = false)
+        private Long episodeId;
+
+        private ViewHistoryId(Long userId, Long episodeId) {
+            this.userId = userId;
+            this.episodeId = episodeId;
+        }
+
+    }
+
     @EmbeddedId
     ViewHistoryId id;
 
     @Builder
     private ViewHistory(Long episodeId, Long userId) {
-        id = ViewHistoryId.builder()
-                .episodeId(episodeId)
-                .userId(userId)
-                .build();
+        id = new ViewHistoryId(episodeId, userId);
     }
-
+    
 }
